+++ app/src/main/java/com/practicum/playlistmaker/creator/Creator.kt
@@ -4,21 +4,21 @@ import android.content.Context
 import android.content.SharedPreferences
 import android.media.MediaPlayer
 import com.practicum.playlistmaker.data.history.HistoryTracksRepositoryImpl
-import com.practicum.playlistmaker.data.mediaPlayer.MediaPlayerRepositoryImpl
-import com.practicum.playlistmaker.data.searchTracks.searchTracksRepositoryImpl
-import com.practicum.playlistmaker.data.searchTracks.network.RetrofitNetworkClient
-import com.practicum.playlistmaker.data.settings.SettingsRepositoryImpl
+import com.practicum.playlistmaker.data.player.MediaPlayerRepositoryImpl
+import com.practicum.playlistmaker.data.search.impl.searchTracksRepositoryImpl
+import com.practicum.playlistmaker.data.search.network.RetrofitNetworkClient
+import com.practicum.playlistmaker.data.settings.impl.SettingsRepositoryImpl
 import com.practicum.playlistmaker.domain.history.api.HistoryTracksInteractor
 import com.practicum.playlistmaker.domain.history.api.HistoryTracksRepository
 import com.practicum.playlistmaker.domain.history.impl.HistoryTracksInteractorImpl
-import com.practicum.playlistmaker.domain.mediaPlayer.api.MediaPlayerInteractor
-import com.practicum.playlistmaker.domain.mediaPlayer.api.MediaPlayerRepository
-import com.practicum.playlistmaker.domain.mediaPlayer.impl.MediaPlayerInteractorImpl
-import com.practicum.playlistmaker.domain.searchTracks.api.SearchTracksInteractor
-import com.practicum.playlistmaker.domain.searchTracks.api.SearchTracksRepository
-import com.practicum.playlistmaker.domain.searchTracks.impl.SearchTracksInteractorImpl
+import com.practicum.playlistmaker.domain.player.api.MediaPlayerInteractor
+import com.practicum.playlistmaker.domain.player.api.MediaPlayerRepository
+import com.practicum.playlistmaker.domain.player.impl.MediaPlayerInteractorImpl
+import com.practicum.playlistmaker.domain.search.api.SearchTracksInteractor
+import com.practicum.playlistmaker.data.search.SearchTracksRepository
+import com.practicum.playlistmaker.domain.search.impl.SearchTracksInteractorImpl
 import com.practicum.playlistmaker.domain.settings.api.SettingsInteractor
-import com.practicum.playlistmaker.domain.settings.api.SettingsRepository
+import com.practicum.playlistmaker.data.settings.SettingsRepository
 import com.practicum.playlistmaker.domain.settings.impl.SettingsInteractorImpl
 import com.practicum.playlistmaker.utils.constants.Constants
