+++ app/src/main/java/com/practicum/playlistmaker/data/search/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.data.search
+
+import com.practicum.playlistmaker.data.search.dto.Response
+
+interface NetworkClient {
+    fun doRequest(dto: Any): Response
+}
