+++ app/src/main/java/com/practicum/playlistmaker/data/search/SearchTracksRepository.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.data.search
+
+import com.practicum.playlistmaker.domain.search.models.SearchTracksResult
+
+interface SearchTracksRepository {
+    fun searchTracks(
+        expression: String
+    ): SearchTracksResult
+}
