+++ app/src/main/java/com/practicum/playlistmaker/data/search/dto/Response.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.data.search.dto
+
+open class Response {
+    var resultCode = 0
+}
