+++ app/src/main/java/com/practicum/playlistmaker/data/search/dto/TracksSearchRequest.kt
@@ -1,0 +1,3 @@
+package com.practicum.playlistmaker.data.search.dto
+
+data class TracksSearchRequest(val expression: String)
