+++ app/src/main/java/com/practicum/playlistmaker/data/search/dto/TracksSearchResponse.kt
@@ -1,0 +1,6 @@
+package com.practicum.playlistmaker.data.search.dto
+
+data class TracksSearchResponse(
+    val resultCount: Int,
+    val results: ArrayList<TrackDto>
+) : Response()
