+++ app/src/main/java/com/practicum/playlistmaker/data/search/network/ITunesApiService.kt
@@ -1,0 +1,15 @@
+package com.practicum.playlistmaker.data.search.network
+
+import com.practicum.playlistmaker.data.search.dto.TracksSearchResponse
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface iTunesApiService {
+
+    @GET("/search?entity=song")
+    fun searchTracks(
+        @Query("term") text: String
+    ): Call<TracksSearchResponse>
+
+}
