--- app/src/main/java/com/practicum/playlistmaker/data/searchTracks/dto/TrackDto.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlistmaker.data.searchTracks.dto
-
-data class TrackDto(val trackId:String,
-                    val trackName: String,
-                    val artistName: String,
-                    val trackTimeMillis: Int,
-                    val artworkUrl100: String,
-                    val collectionName: String,
-                    val releaseDate: String,
-                    val primaryGenreName: String,
-                    val country: String,
-                    val previewUrl: String)
