--- app/src/main/java/com/practicum/playlistmaker/data/searchTracks/dto/TracksSearchRequest.kt
@@ -1,3 +1,0 @@
-package com.practicum.playlistmaker.data.searchTracks.dto
-
-data class TracksSearchRequest(val expression: String)
