--- app/src/main/java/com/practicum/playlistmaker/data/searchTracks/dto/TracksSearchResponse.kt
@@ -1,6 +1,0 @@
-package com.practicum.playlistmaker.data.searchTracks.dto
-
-data class TracksSearchResponse(
-    val resultCount: Int,
-    val results: ArrayList<TrackDto>
-) : Response()
