+++ app/src/main/java/com/practicum/playlistmaker/data/settings/SettingsRepository.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker.data.settings
+
+import android.content.Context
+
+interface SettingsRepository {
+    fun isThemeInSettingsDark(context: Context): Boolean
+    fun writeThemeInSettingsDark(setDark: Boolean)
+}
