--- app/src/main/java/com/practicum/playlistmaker/domain/mediaPlayer/api/MediaPlayerRepository.kt
@@ -1,15 +1,0 @@
-package com.practicum.playlistmaker.domain.mediaPlayer.api
-
-interface MediaPlayerRepository {
-    fun prepare(
-        url: String?,
-        onPrepared: () -> Unit,
-        onCompletion: () -> Unit
-    )
-    fun start()
-    fun pause()
-    fun release()
-    fun timerUpdate(
-        onTimerUpdated: (String) -> Unit
-    )
-}
