--- app/src/main/java/com/practicum/playlistmaker/domain/mediaPlayer/models/PlayerStatus.kt
@@ -1,8 +1,0 @@
-package com.practicum.playlistmaker.domain.mediaPlayer.models
-
-enum class PlayerStatus {
-    STATE_DEFAULT,
-    STATE_PREPARED,
-    STATE_PLAYING,
-    STATE_PAUSED
-}
