+++ app/src/main/java/com/practicum/playlistmaker/domain/player/api/MediaPlayerRepository.kt
@@ -1,0 +1,15 @@
+package com.practicum.playlistmaker.domain.player.api
+
+interface MediaPlayerRepository {
+    fun prepare(
+        url: String?,
+        onPrepared: () -> Unit,
+        onCompletion: () -> Unit
+    )
+    fun start()
+    fun pause()
+    fun release()
+    fun timerUpdate(
+        onTimerUpdated: (String) -> Unit
+    )
+}
