+++ app/src/main/java/com/practicum/playlistmaker/domain/player/models/PlayerStatus.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker.domain.player.models
+
+enum class PlayerStatus {
+    STATE_DEFAULT,
+    STATE_PREPARED,
+    STATE_PLAYING,
+    STATE_PAUSED
+}
