+++ app/src/main/java/com/practicum/playlistmaker/domain/search/api/SearchTracksInteractor.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.domain.search.api
+
+import com.practicum.playlistmaker.domain.search.models.SearchTracksResult
+
+interface SearchTracksInteractor {
+    fun searchTracks(expression: String, consumer: TracksConsumer)
+
+    interface TracksConsumer {
+        fun consume(result: SearchTracksResult)
+    }
+}
