+++ app/src/main/java/com/practicum/playlistmaker/domain/search/impl/SearchTracksInteractorImpl.kt
@@ -1,0 +1,17 @@
+package com.practicum.playlistmaker.domain.search.impl
+
+import com.practicum.playlistmaker.domain.search.api.SearchTracksInteractor
+import com.practicum.playlistmaker.data.search.SearchTracksRepository
+import java.util.concurrent.Executors
+
+class SearchTracksInteractorImpl(private val repository: SearchTracksRepository): SearchTracksInteractor {
+
+    private val executor = Executors.newCachedThreadPool()
+
+    override fun searchTracks(expression: String, consumer: SearchTracksInteractor.TracksConsumer) {
+
+        executor.execute {
+            consumer.consume(repository.searchTracks(expression))
+        }
+    }
+}
