+++ app/src/main/java/com/practicum/playlistmaker/domain/search/models/SearchStatus.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.domain.search.models
+
+enum class SearchStatus {
+    DEFAULT,
+    SEARCH_RESULT_WAITING,
+    HISTORY_PLACEHOLDER,
+    TRACKS_FOUND,
+    TRACKS_NOT_FOUND,
+    SOMETHING_WRONG,
+    ERROR_OCCURRED
+}
