--- app/src/main/java/com/practicum/playlistmaker/domain/searchTracks/api/SearchTracksInteractor.kt
@@ -1,11 +1,0 @@
-package com.practicum.playlistmaker.domain.searchTracks.api
-
-import com.practicum.playlistmaker.domain.searchTracks.models.SearchTracksResult
-
-interface SearchTracksInteractor {
-    fun searchTracks(expression: String, consumer: TracksConsumer)
-
-    interface TracksConsumer {
-        fun consume(result: SearchTracksResult)
-    }
-}
