--- app/src/main/java/com/practicum/playlistmaker/domain/searchTracks/api/SearchTracksRepository.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmaker.domain.searchTracks.api
-
-import com.practicum.playlistmaker.domain.searchTracks.models.SearchTracksResult
-
-interface SearchTracksRepository {
-    fun searchTracks(
-        expression: String
-    ): SearchTracksResult
-}
