--- app/src/main/java/com/practicum/playlistmaker/domain/settings/api/SettingsRepository.kt
@@ -1,8 +1,0 @@
-package com.practicum.playlistmaker.domain.settings.api
-
-import android.content.Context
-
-interface SettingsRepository {
-    fun isThemeInSettingsDark(context: Context): Boolean
-    fun writeThemeInSettingsDark(setDark: Boolean)
-}
