+++ app/src/main/java/com/practicum/playlistmaker/domain/settings/impl/SettingsInteractorImpl.kt
@@ -1,9 +1,8 @@
 package com.practicum.playlistmaker.domain.settings.impl
 
 import android.content.Context
-import com.practicum.playlistmaker.data.settings.SettingsRepositoryImpl
 import com.practicum.playlistmaker.domain.settings.api.SettingsInteractor
-import com.practicum.playlistmaker.domain.settings.api.SettingsRepository
+import com.practicum.playlistmaker.data.settings.SettingsRepository
 
 class SettingsInteractorImpl(private val settingsRepository: SettingsRepository): SettingsInteractor {
     override fun isThemeDark(context: Context): Boolean {
