+++ app/src/main/java/com/practicum/playlistmaker/ui/search/TrackViewHolder.kt
@@ -7,7 +7,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.domain.searchTracks.models.Track
+import com.practicum.playlistmaker.domain.search.models.Track
 import com.practicum.playlistmaker.utils.converters.dimensionsFloatToIntConvert
 import java.text.SimpleDateFormat
 import java.util.Locale
